Refactor MCP transform tool
